chore(swap-verification): rename Maya -> Mayachain for naming consistency
Align the MAYAChain verifier naming with the fully-spelled Thorchain
convention: verifyMaya -> verifyMayachain, test fixtures maya/ -> mayachain/
(mayaResponse/mayaSwap -> mayachainResponse/mayachainSwap), maya.test.ts ->
mayachain.test.ts, and the RUNE/CACAO precision comment. No behavior change.
